+++ src/main/java/com/codeborne/selenide/ex/ElementIsNotClickableException.java
@@ -4,7 +4,7 @@ import javax.annotation.ParametersAreNonnullByDefault;
 
 @ParametersAreNonnullByDefault
 public class ElementIsNotClickableException extends UIAssertionError {
-  public ElementIsNotClickableException(Throwable cause) {
-    super("Element is not clickable", cause);
+  public ElementIsNotClickableException(String elementDescription, Throwable cause) {
+    super("Element is not clickable: " + elementDescription, cause);
   }
 }