+++ src/test/java/com/codeborne/selenide/ex/InvalidStateExceptionTest.java
@@ -9,11 +9,11 @@ final class InvalidStateExceptionTest {
   @Test
   void constructorWithCause() {
     StaleElementReferenceException cause = new StaleElementReferenceException("Houston, we have a problem");
-    InvalidStateException invalidStateException = new InvalidStateException(cause);
+    InvalidStateException invalidStateException = new InvalidStateException("#link", cause);
 
     assertThat(invalidStateException).hasMessageStartingWith("Invalid element state");
     assertThat(invalidStateException).hasMessageEndingWith("StaleElementReferenceException: Houston, we have a problem");
-    assertThat(invalidStateException).hasToString(String.format("Invalid element state: " +
+    assertThat(invalidStateException).hasToString(String.format("Invalid element state [#link]: " +
       "Houston, we have a problem%n" +
       "Timeout: 0 ms.%n" +
       "Caused by: StaleElementReferenceException: Houston, we have a problem"));
@@ -21,10 +21,10 @@ final class InvalidStateExceptionTest {
 
   @Test
   void constructorWithMessage() {
-    InvalidStateException invalidStateException = new InvalidStateException("Houston, we have a problem");
+    InvalidStateException invalidStateException = new InvalidStateException("#link", "Houston, we have a problem");
 
-    assertThat(invalidStateException).hasMessageStartingWith("Invalid element state: Houston, we have a problem");
-    assertThat(invalidStateException).hasToString(String.format("Invalid element state: Houston, we have a problem%n" +
+    assertThat(invalidStateException).hasMessageStartingWith("Invalid element state [#link]: Houston, we have a problem");
+    assertThat(invalidStateException).hasToString(String.format("Invalid element state [#link]: Houston, we have a problem%n" +
       "Timeout: 0 ms."));
   }
 }