+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 6.2.0
 * #1589 add "<Click to see difference>" to most of Selenide assertion errors
+* add locator to some of Selenide error messages
 
 ## 6.1.2 (released 22.12.2021)
 * #1672 upgrade to Selenium 4.1.1